share MLP implementations used 4+ times
Add mlx_vlm/models/mlp.py with the MLP variants that appear 4 or more times across
models (SwiGLUMLP, DeepseekMLP, GELUMLP, FastGELUMLP, TanhGELUMLP) and repoint the
46 duplicate inline definitions to import them. Behavior-preserving: each shared
class is identical to the originals it replaces; the SwiGLU spellings are unified
via the existing swiglu helper (== silu(gate) * up).
