+++ mlx_vlm/models/deepseek_vl_v2/language.py
@@ -9,6 +9,7 @@ from ..base import (
     create_attention_mask,
     scaled_dot_product_attention,
 )
+from ..mlp import DeepseekMLP as DeepseekV2MLP
 from ..switch_layers import SwitchGLU
 from .config import TextConfig
 
@@ -283,26 +284,6 @@ class LlamaAttention(nn.Module):
 
         output = output.transpose(0, 2, 1, 3).reshape(B, L, -1)
         return self.o_proj(output)
-
-
-class DeepseekV2MLP(nn.Module):
-    def __init__(
-        self, config: TextConfig, hidden_size: int = None, intermediate_size: int = None
-    ):
-        super().__init__()
-        self.config = config
-        self.hidden_size = config.hidden_size if hidden_size is None else hidden_size
-        self.intermediate_size = (
-            config.intermediate_size if intermediate_size is None else intermediate_size
-        )
-
-        self.gate_proj = nn.Linear(self.hidden_size, self.intermediate_size, bias=False)
-        self.up_proj = nn.Linear(self.hidden_size, self.intermediate_size, bias=False)
-        self.down_proj = nn.Linear(self.intermediate_size, self.hidden_size, bias=False)
-
-    def __call__(self, x):
-        down_proj = self.down_proj(nn.silu(self.gate_proj(x)) * self.up_proj(x))
-        return down_proj
 
 
 class MoEGate(nn.Module):
