+++ mlx_vlm/models/glm4v/vision.py
@@ -4,6 +4,7 @@ import mlx.core as mx
 import mlx.nn as nn
 
 from ..kernels import grid_sample
+from ..mlp import SwiGLUMLP as Glm4vVisionMLP
 from .config import VisionConfig
 
 
@@ -224,17 +225,6 @@ class Glm4vVisionAttention(nn.Module):
         output = mx.concatenate(attn_outputs, axis=2)
         output = output.transpose(0, 2, 1, 3).reshape(seq_length, -1)
         return self.proj(output)
-
-
-class Glm4vVisionMLP(nn.Module):
-    def __init__(self, dim, hidden_dim):
-        super().__init__()
-        self.gate_proj = nn.Linear(dim, hidden_dim, bias=False)
-        self.up_proj = nn.Linear(dim, hidden_dim, bias=False)
-        self.down_proj = nn.Linear(hidden_dim, dim, bias=False)
-
-    def __call__(self, x: mx.array) -> mx.array:
-        return self.down_proj(nn.silu(self.gate_proj(x)) * self.up_proj(x))
 
 
 class Glm4vVisionBlock(nn.Module):
