+++ mlx_vlm/models/moondream3/language.py
@@ -5,6 +5,7 @@ import mlx.nn as nn
 
 from ..base import LanguageModelOutput, create_attention_mask
 from ..cache import KVCache
+from ..mlp import TanhGELUMLP as DenseMLP
 from ..switch_layers import SwitchLinear
 from .config import TextConfig
 
@@ -149,17 +150,6 @@ class Attention(nn.Module):
         )
         output = output.transpose(0, 2, 1, 3).reshape(B, L, -1)
         return self.proj(output)
-
-
-class DenseMLP(nn.Module):
-    def __init__(self, config: TextConfig):
-        super().__init__()
-        self.fc1 = nn.Linear(config.hidden_size, config.intermediate_size, bias=True)
-        self.fc2 = nn.Linear(config.intermediate_size, config.hidden_size, bias=True)
-        self.act = nn.GELU(approx="tanh")
-
-    def __call__(self, x: mx.array) -> mx.array:
-        return self.fc2(self.act(self.fc1(x)))
 
 
 @mx.compile
